Next stable version. Add Asp.Core.Identity. Implement ReturnProtocolBuildersStore, ReturnProtocolBuilderService and 2 new ReturnProtocolBuilders. Implement WPF Login View and logic to create users in Asp.Core.Identity. Remove Unusing classes.

+++ Components/ComponentsViewModel/ReturnProtocolComponentViewModels/ListViewInputViewModel.cs
@@ -5,6 +5,8 @@ using Microsoft.Extensions.DependencyInjection;
 using System.Collections.ObjectModel;
 using DelitaTrade.Core.Contracts;
 using DelitaTrade.Core.ViewModels;
+using DelitaTrade.ViewModels.Controllers;
+using System.Threading.Tasks;
 
 namespace DelitaTrade.Components.ComponentsViewModel.ReturnProtocolComponentViewModels
 {
@@ -15,29 +17,27 @@ namespace DelitaTrade.Components.ComponentsViewModel.ReturnProtocolComponentView
         private ProductViewModel _selectedProduct;
         private ReturnProtocolController _returnProtocolController;
         private ReturnProtocolViewModel? _currentReturnProtocolViewModel;
+        private readonly DescriptionCategoryController _descriptionCategoryController;
         
         private readonly ObservableCollection<ProductToReturnViewModel> _list;
         private ObservableCollection<string> _productUnit;
-        private ObservableCollection<ProductViewModel> _products = new ObservableCollection<ProductViewModel>();
-        private ObservableCollection<ReturnedProductDescriptionViewModel> _description = new ObservableCollection<ReturnedProductDescriptionViewModel>();
 
-        public ListViewInputViewModel(IServiceProvider serviceProvider, ReturnProtocolController returnProtocolViewModel)
+        public ListViewInputViewModel(IServiceProvider serviceProvider, ReturnProtocolController returnProtocolViewModel, DescriptionCategoryController descriptionCategoryController)
         {
             _serviceProvider = serviceProvider;
             _returnProtocolController = returnProtocolViewModel;
+            _descriptionCategoryController = descriptionCategoryController;
             _list = new ObservableCollection<ProductToReturnViewModel>();
-            _productUnit = new ObservableCollection<string> { ProductUnit.Count, ProductUnit.Kg, ProductUnit.Box };
+            _productUnit = new ObservableCollection<string> { ProductUnit.Count, ProductUnit.Kg, ProductUnit.Box, ProductUnit.Liter, ProductUnit.Pack, ProductUnit.Meter };
             ProductCreate += OnCreatedProduct;
-            DescriptionCreate += OnDescriptionCreate;
-            UpdateProduct();
+            DescriptionCreate += OnCreateDescription;
             _returnProtocolController.ReturnProtocolSelected += InitializedList;
             _returnProtocolController.ReturnProtocolUnSelected += UnselectedProtocol;
             ReturnedProductCreate += AddReturnedProduct;
         }
 
         public ObservableCollection<ProductToReturnViewModel> List => _list;
-        public ObservableCollection<ProductViewModel> ProductsList => _products;
-        public ObservableCollection<ReturnedProductDescriptionViewModel> Descriptions => _description;
+        public DescriptionCategoryController DescriptionCategoryController => _descriptionCategoryController;
         public ObservableCollection<string> Unit => _productUnit;
         public event Func<ProductViewModel, Task> ProductCreate;
         public event Func<ReturnedProductDescriptionViewModel, Task<ReturnedProductDescriptionViewModel>> DescriptionCreate;
@@ -54,9 +54,11 @@ namespace DelitaTrade.Components.ComponentsViewModel.ReturnProtocolComponentView
 
         public async void RemoveRow(ProductToReturnViewModel returnedProduct) 
         {
+            if (_currentReturnProtocolViewModel == null) throw new ArgumentNullException("No return protocol loaded");
             using var scope = _serviceProvider.CreateScope();
             var service = scope.GetService<IReturnProductService>();
             await service.DeleteProductAsync(returnedProduct.Id);
+            _currentReturnProtocolViewModel.RemoveProductById(returnedProduct.Id);
             _list.Remove(returnedProduct);
         }
 
@@ -70,32 +72,77 @@ namespace DelitaTrade.Components.ComponentsViewModel.ReturnProtocolComponentView
             return returnProduct.Id;
         }
 
+        private async Task UpdateRow(ReturnedProductViewModel returnedProduct)
+        {
+            if (_currentReturnProtocolViewModel == null) throw new ArgumentNullException("No return protocol loaded");
+            using var scope = _serviceProvider.CreateScope();
+            var service = scope.GetService<IReturnProductService>();
+            await service.UpdateProductAsync(returnedProduct);
+        }
+
         private async void AddRow()
         {
             if (_list.Count > 0)
             {
-                var product = new ProductViewModel { Name = _list[^1].ProductName, Unit = _list[^1].Unit };
-                var description = new ReturnedProductDescriptionViewModel(_list[^1].Description);
-                _list[^1].IsCompleted -= AddRow;                
+                var product = new ProductViewModel { Name = _list[^1].ProductName, Unit = _list[^1].Unit, Number = _list[^1].Number };
+                ReturnedProductDescriptionViewModel? description = null;
+                ReturnedProductDescriptionViewModel? resultDescription = null;
+                if (string.IsNullOrEmpty(_list[^1].Description) == false)
+                {
+                    description = new ReturnedProductDescriptionViewModel(_list[^1].Description!);
+                    resultDescription = await DescriptionCreate(description);
+                }
+                _list[^1].IsCompleted -= AddRow;
                 var t = ProductCreate(product);
-                description = await DescriptionCreate(description);
                 await t;
-                _list[^1].Id = await ReturnedProductCreate(new ReturnedProductViewModel
+                ReturnedProductViewModel newProduct = new ReturnedProductViewModel
                 {
                     Batch = _list[^1].Batch,
-                    Quantity = _list[^1].ProductQuantity,
-                    BestBefore = _list[^1].BestBefore,
+                    Quantity = _list[^1].ProductQuantity > 0 ? _list[^1].ProductQuantity : throw new ArgumentException("Product quantity must be greater than 0"),
+                    BestBefore = _list[^1].BestBefore ?? throw new ArgumentNullException("Best before property is required"),
                     Product = product,
-                    Description = description
-                });
-                AddProduct(product);
-                AddDescription(description);
+                    Description = resultDescription,
+                    DescriptionCategory = _list[^1].DescriptionCategory ?? throw new ArgumentNullException("Description category is required"),
+                };
+                newProduct.Id = await ReturnedProductCreate(newProduct);
+                _list[^1].Id = newProduct.Id;
+                _list[^1].SetCreatedProduct(newProduct);
+                _list[^1].UpdateProduct += OnProductUpdate;
             }
-            _list.Add(new("", this));
+            _list.Add(new(this, _serviceProvider.GetRequiredService<ProductSearchController>()));
             _list[^1].IsCompleted += AddRow;            
         }
+
+        private async void OnProductUpdate(ProductToReturnViewModel model)
+        {
+            var productToUpdate = _currentReturnProtocolViewModel!.Products.FirstOrDefault(x => x.Id == model.Id);
+            if (productToUpdate == null) throw new ArgumentNullException("Product to update not found");
+
+            var product = new ProductViewModel { Name = model.ProductName, Unit = model.Unit, Number = model.Number };
+            ReturnedProductDescriptionViewModel? description = null;
+            ReturnedProductDescriptionViewModel? resultDescription = null;
+            if (string.IsNullOrWhiteSpace(model.Description) == false)
+            {
+                description = new ReturnedProductDescriptionViewModel(model.Description);
+                resultDescription = await DescriptionCreate(description);
+            }
+            var t = ProductCreate(product);
+            await t;
+
+            productToUpdate.Batch = model.Batch;
+            productToUpdate.Quantity = model.ProductQuantity;
+            productToUpdate.BestBefore = model.BestBefore ?? throw new ArgumentNullException("Best before date is required");
+            productToUpdate.Product = product;
+            productToUpdate.Description = resultDescription;
+            productToUpdate.DescriptionCategory = model.DescriptionCategory ?? throw new ArgumentNullException("Description category is required");
+
+            await UpdateRow(productToUpdate);
+        }
+
         private void InitializedList(ReturnProtocolViewModel returnProtocol)
         {
+            UnselectedProtocol();
+
             if (returnProtocol.Products.Count > 0)
             {
                 LoadList(returnProtocol);
@@ -110,7 +157,7 @@ namespace DelitaTrade.Components.ComponentsViewModel.ReturnProtocolComponentView
         {
             _currentReturnProtocolViewModel = returnProtocol;
             _list.Clear();
-            _list.Add(new("", this));
+            _list.Add(new(this, _serviceProvider.GetRequiredService<ProductSearchController>()));
             _list[^1].IsCompleted += AddRow;
         }
 
@@ -120,44 +167,11 @@ namespace DelitaTrade.Components.ComponentsViewModel.ReturnProtocolComponentView
             _list.Clear();
             foreach (var item in returnProtocol.Products)
             {
-                _list.Add(new ProductToReturnViewModel(item, this));
+                _list.Add(new ProductToReturnViewModel(item, this, _serviceProvider.GetRequiredService<ProductSearchController>()));
+                _list[^1].UpdateProduct += OnProductUpdate;
             }
-            _list.Add(new("", this));
+            _list.Add(new(this, _serviceProvider.GetRequiredService<ProductSearchController>()));
             _list[^1].IsCompleted += AddRow;
-        }
-
-        private void AddProduct(ProductViewModel product) 
-        {
-            if (_products.FirstOrDefault(p => p.Name == product.Name && p.Unit == product.Unit) == null)
-            { 
-                _products.Add(product);
-            }
-        }
-
-        private void AddDescription(ReturnedProductDescriptionViewModel description)
-        {
-            if (_description.FirstOrDefault(d => d.Description == description.Description) == null)
-            { 
-                _description.Add(description);
-            }
-        }
-
-        private async void UpdateProduct()
-        {
-            using var scope = _serviceProvider.CreateScope();
-            var products = await scope.GetService<IProductService>().GetAllAsync();
-            _products.Clear();
-            foreach (var prod in products)
-            {
-                _products.Add(prod);
-            }                
-            
-            var descriptions = await scope.GetService<IProductDescriptionService>().GetAllAsync();
-            _description.Clear();
-            foreach (var desc in descriptions)
-            {
-                _description.Add(desc);
-            }
         }
 
         private async Task OnCreatedProduct(ProductViewModel product)
@@ -167,17 +181,22 @@ namespace DelitaTrade.Components.ComponentsViewModel.ReturnProtocolComponentView
             await service.AddProductAsync(product);
         }
 
-        private async Task<ReturnedProductDescriptionViewModel> OnDescriptionCreate(ReturnedProductDescriptionViewModel description)
+        private async Task<ReturnedProductDescriptionViewModel> OnCreateDescription(ReturnedProductDescriptionViewModel description)
         {
             using var scope = _serviceProvider.CreateScope();
             var service = scope.GetService<IProductDescriptionService>();
-            return await service.AddDescription(description);
+            return await service.AddDescriptionAsync(description);
         }
 
         private void UnselectedProtocol()
         {
             if (_list.Count > 0)
             {
+                for (int i = 0; i < _list.Count - 1; i++)
+                {
+                    _list[i].UpdateProduct -= OnProductUpdate;
+                }
+
                 _list[^1].IsCompleted -= AddRow;
                 _list.Clear();
             }
